feat: add useUploadFiles hook to replace the repeat logic code
