+++ src/features/PageEditor/Copilot/Conversation.tsx
@@ -1,14 +1,12 @@
 import { Flexbox } from '@lobehub/ui';
-import { memo, useCallback, useEffect, useState } from 'react';
+import { memo, useEffect, useState } from 'react';
 
-import DragUploadZone from '@/components/DragUploadZone';
+import DragUploadZone, { useUploadFiles } from '@/components/DragUploadZone';
 import type { ActionKeys } from '@/features/ChatInput';
 import { ChatInput, ChatList } from '@/features/Conversation';
-import { useModelSupportVision } from '@/hooks/useModelSupportVision';
 import { useAgentStore } from '@/store/agent';
 import { agentByIdSelectors } from '@/store/agent/selectors';
 import { useChatStore } from '@/store/chat';
-import { useFileStore } from '@/store/file';
 
 import CopilotToolbar from './Toolbar';
 import Welcome from './Welcome';
@@ -38,22 +36,7 @@ const Conversation = memo<ConversationProps>(({ agentId }) => {
   const provider = useAgentStore((s) =>
     agentByIdSelectors.getAgentModelProviderById(currentAgentId)(s),
   );
-  const canUploadImage = useModelSupportVision(model, provider);
-  const uploadFiles = useFileStore((s) => s.uploadChatFiles);
-
-  const handleUploadFiles = useCallback(
-    async (files: File[]) => {
-      const filteredFiles = files.filter((file) => {
-        if (canUploadImage) return true;
-        return !file.type.startsWith('image');
-      });
-
-      if (filteredFiles.length > 0) {
-        uploadFiles(filteredFiles);
-      }
-    },
-    [canUploadImage, uploadFiles],
-  );
+  const { handleUploadFiles } = useUploadFiles({ model, provider });
 
   return (
     <DragUploadZone
